job_matcher/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
pypdf>=4.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
chromadb>=0.5.0
